requirements/testing.txt
-r optional.txt
mock
pytest==3.1.3 ; python_version < '3.4'
pytest ; python_version > '3.3'
pytest-xdist
pytest-html
pytest-django
pytest-sugar
pytest-cov
python-coveralls
factory-boy
psycopg2
django-environ
pathlib ; python_version < '3.3'
enum34 ; python_version < '3.4'
